+++ src/main/java/team7/hrbank/domain/employee/entity/EmployeeStatus.java
@@ -1,4 +1,4 @@
-package team7.hrbank.domain.employee;
+package team7.hrbank.domain.employee.entity;
 
 public enum EmployeeStatus {
     ACTIVE,ON_LEAVE, RESIGNED
